requirements-test.txt
# Test dependencies for Frigate Config Builder
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pyyaml>=6.0
